Split out week extraction and date parsing to a separate file
